test(encoder): implement Byte data encoding

+++ crates/qr_encoder/src/data_encoders/data_encoder.rs
@@ -1,5 +1,5 @@
 use bitvec::{order::Msb0, vec::BitVec};
 
 pub trait DataEncoder {
-    fn encode(&self, data: &str) -> BitVec<usize, Msb0>;
+    fn encode(&self, data: &str) -> BitVec<u8, Msb0>;
 }
